% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_realAlzheimersResults.R
\docType{data}
\name{alzheimersMultiOmics_df}
\alias{alzheimersMultiOmics_df}
\title{Real Multi-Omics Data Results for AD}
\format{
A tibble (a modern data frame created by the \code{tibble} package)
with eight columns and 2833 rows:
\itemize{
\item{\code{pathway}}{ the name of the pathway / gene set in C2 CP}
\item{\code{size}}{ the number of genes in the pathway / gene set}
\item{\code{snpPval}}{ the \emph{p}-values for each pathway after SNP GWAS
analysis}
\item{\code{snpFDR}}{ the false discovery rate for each SNP \emph{p}-value}
\item{\code{dnamPval}}{ the \emph{p}-values for each pathway after DNA
methylation analysis with \code{missMethyl::}}
\item{\code{dnamFDR}}{ the false discovery rate for each DNAm \emph{p}-value}
\item{\code{rnaseqPval}}{ the \emph{p}-values for each pathway after RNAseq
analysis with \code{fgsea::}}
\item{\code{rnaseqFDR}}{ the false discovery rate for each RNAseq \emph{p}-value}
}
}
\usage{
alzheimersMultiOmics_df
}
\description{
Real multi-omics data results for Alzheimer's disease using SNP,
DNA methylation, and gene expression data from the ROSMAP study
}
\details{
This data gives pathway-level statistical significance results from
GSEA under Bacon correction (for SNP analysis), \code{missMethyl} (for DNA
methylation analysis), and \code{fgsea} (for gene expression analysis) for
640 samples from the ROSMAP study using the Broad Institute's C2 CP
collection (2833 pathways with the number of genes less than 200 or more
than 4).
}
\keyword{datasets}
