% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MiniMax_drivers.R
\name{MiniMax_calculateDrivers}
\alias{MiniMax_calculateDrivers}
\title{Mark which Platforms are Driving the MiniMax Statistics}
\usage{
MiniMax_calculateDrivers(
  res_df,
  orderStat = 2L,
  drivers_char = colnames(res_df),
  sortLabels = TRUE,
  separator = " and "
)
}
\arguments{
\item{res_df}{A data frame of \emph{p}-values. The rows correspond to gene
sets / pathways and the columns correspond to a data platform for the
disease of interest.}

\item{orderStat}{How many platforms should show a biological signal for a
pathway / gene set to have multi-omic "enrichment"? Defaults to 2. See
"Details" for more information.}

\item{drivers_char}{What labels should be given to the driving platforms?
Defaults to the column names of \code{res_df}. If you supply custom labels,
make sure to match them to the column order of \code{res_df}.}

\item{sortLabels}{Should the driver labels be sorted alphabetically before
concatenation? Defaults to \code{TRUE}; that is, a multi-omics result
driven first by protein expression then by DNA methylation will have the
same label as a result driven first by DNA methylation then by protein
expression. If you would like the magnitude of the \emph{p}-value to set
the label order, then use \code{sortLabels = FALSE}.}

\item{separator}{What character string should be used to separate the names
of the driving platforms? Defaults to \code{" and "}; for example, if the
platform driver labels are \code{"protein"} and \code{"cnv"}, and if
\code{sortLabels = TRUE}, then the label of drivers would be
\code{"cnv and protein"}.}
}
\value{
A vector of the names of the platforms driving the MiniMax statistic
values.
}
\description{
Given a data frame of pathway-level \emph{p}-values, mark which
of the multi-omics platforms are driving the MiniMax statistic.
}
\details{
The MiniMax statistic is defined as the minimum of all pairwise
maxima of pathway \emph{p}-values. This operation is arithmetically
equivalent to sorting the \emph{p}-values and taking the second smallest.
In our experience, setting this "order statistic" cutoff to 2 is
appropriate for =< 5 data platforms. Biologically, this is equivalent to
saying "if this pathway is dysregulated in at least two data types for
this disease / condition, it is worthy of additional consideration". In
situations where more than 5 data platforms are available for the disease
of interest, we recommend increasing the \code{orderStat} value to 3.

NOTE: this result does not depend on the pathway significance level at all.
This result will simply show you which platforms had the smallest
\emph{p}-values for a particular pathway, even if the MiniMax statistic is
not statistically significant for that pathway. Therefore, we recommend
that this function be used only for interpretation of results post-hoc.
}
\examples{
  MiniMax_calculateDrivers(
    multiOmicsHighSignalResults_df[, -(1:2)],
    drivers_char = c("cnv", "rnaSeq", "protein")
 )

}
