% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MiniMax_significance.R
\name{MiniMax_calculatePVal}
\alias{MiniMax_calculatePVal}
\title{Calculate the MiniMax \emph{p}-Values}
\usage{
MiniMax_calculatePVal(MiniMax_num, betaParams_ls)
}
\arguments{
\item{MiniMax_num}{A numeric vector of MiniMax statistics}

\item{betaParams_ls}{A list of the parameters for the Beta Distribution.
These values should be returned by the \code{\link{MiniMax_estBetaParams}}
function.}
}
\value{
A vector of the MiniMax significance levels (\emph{p}-values)
corresponding to the supplied MiniMax statistics.
}
\description{
Given a vector of pathway-level MiniMax statistics and a vector
of Beta Distribution parameters, calculate the MiniMax statistics'
\emph{p}-values.
}
\examples{
  # Find the best-fitting paramters based on the MiniMax statistic values
  #   under the null distribution
  mmBetaParams_ls <- MiniMax_estBetaParams(
    MiniMaxNull_num = nullMiniMaxResults_df$MiniMax,
    nPlatforms = 3L,
    method = "MoM"
  )

  # Calculate the MiniMax Statistics for each gene set
  mmVals_num <- MiniMax_calculateStatistic(
    res_df = multiOmicsHighSignalResults_df[, -(1:2)]
  )

  # Find the p-values corresponding to these statistics
  MiniMax_calculatePVal(
    MiniMax_num = mmVals_num,
    betaParams_ls = mmBetaParams_ls
 )

}
