% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MiniMax_statistic.R
\name{MiniMax_calculateStatistic}
\alias{MiniMax_calculateStatistic}
\title{Calculate the MiniMax Statistics}
\usage{
MiniMax_calculateStatistic(res_df, orderStat = 2L)
}
\arguments{
\item{res_df}{A data frame of \emph{p}-values. The rows correspond to gene
sets / pathways and the columns correspond to a data platform for the
disease of interest.}

\item{orderStat}{How many platforms should show a biological signal for a
pathway / gene set to have multi-omic "enrichment"? Defaults to 2. See
"Details" for more information.}
}
\value{
A vector of the MiniMax statistic values.
}
\description{
Given a data frame of pathway-level \emph{p}-values, calculate
the MiniMax statistic (but not the significance level of this statistic).
}
\details{
The MiniMax statistic is defined as the minimum of all pairwise
maxima of pathway \emph{p}-values. This operation is arithmetically
equivalent to sorting the \emph{p}-values and taking the second smallest.
In our experience, setting this "order statistic" cutoff to 2 is
appropriate for =< 5 data platforms. Biologically, this is equivalent to
saying "if this pathway is dysregulated in at least two data types for
this disease / condition, it is worthy of additional consideration". In
situations where more than 5 data platforms are available for the disease
of interest, we recommend increasing the \code{orderStat} value to 3.
}
\examples{
  MiniMax_calculateStatistic(multiOmicsHighSignalResults_df[, -(1:2)])

}
