% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_MiniMax_params.R
\name{MiniMax_estBetaParams}
\alias{MiniMax_estBetaParams}
\title{Estimate the Parameters of the MiniMax Statistic's Beta Distribution}
\usage{
MiniMax_estBetaParams(
  MiniMaxNull_num,
  nPlatforms,
  orderStat = 2L,
  method = c("parametric", "MLE", "MoM")
)
}
\arguments{
\item{MiniMaxNull_num}{A numeric vector of MiniMax statistics under the null}

\item{nPlatforms}{An integer stating how many data platforms are in the
original data.}

\item{orderStat}{How many platforms should show a biological signal for a
pathway / gene set to have multi-omic "enrichment"? Defaults to 2. See
"Details" for more information.}

\item{method}{Which estimation method will be used to find the parameters of
the Beta Distribution? Options are \code{"parametric"} (no estimation from
the data), \code{"MLE"} (Maximum Likelihood Estimates), or \code{"MoM"}
(Method of Moments estimates). See "Details" for more information.}
}
\value{
A list of 3 components: "alpha" and "beta" hold the parameter
estimates of the Beta Distribution, and "method" returns a character
string denoting which estimation method was used.
}
\description{
Given a vector of MiniMax statisic values under the null
hypothesis, estimate the parameters of the Beta Distribution which best
fits these values.
}
\details{
\strong{Concerning Parameter Estimation Methods:} We currently support 3
options to estimate the parameters of the Beta Distribution. The
"parametric" option does not use the data. Instead, it assumes that the
MiniMax statistics will have a Beta \eqn{(k, n + 1 - k)} distribution,
where \eqn{k} is the value of \code{orderStat} and \eqn{n} has the value
\code{nPlatforms}. See \url{https://en.wikipedia.org/wiki/Order_statistic}.

The next two estimation options make use of the \code{MiniMaxNull_num}
vector, which should be calculated by finding the same significance levels
of the statistical tests used on the real data (for each pathway and data
platform), but by using a random permutation of the outcome of interest
instead of the real values; more permutations are better. The "MLE" option
uses the \code{\link[Rfast]{beta.mle}} function to find the Maximum
Likelihood Estimates of \eqn{\alpha} and \eqn{\beta}. The "MoM" option uses
the closed-form Method of Moments estimators of \eqn{\alpha} and
\eqn{\beta} as shown in
\url{https://en.wikipedia.org/wiki/Beta_distribution#Method_of_moments}.

\strong{Concerning Appropriate Order Statistics:} The MiniMax operation is
equivalent to sorting the \emph{p}-values and taking the second smallest.
In our experience, setting this "order statistic" cutoff to 2 is
appropriate for =< 5 data platforms. Biologically, this is equivalent to
saying "if this pathway is dysregulated in at least two data types for
this disease / condition, it is worthy of additional consideration". In
situations where more than 5 data platforms are available for the disease
of interest, we recommend increasing the \code{orderStat} value to 3.
}
\examples{
 miniMax_num <- nullMiniMaxResults_df$MiniMax

 MiniMax_estBetaParams(miniMax_num, nPlatforms = 3L)
 MiniMax_estBetaParams(miniMax_num, nPlatforms = 3L, method = "MoM")
 MiniMax_estBetaParams(miniMax_num, nPlatforms = 3L, method = "MLE")

}
